GEODE-7601: Memory leaks in C++ native client found when running old integration-tests (#564)
Several memory leaks and other issues in memory
management have been fixed in the native client.
The issues were found when running the old integration
tests under valgrind.

Many issues in the testing framework have also been fixed
to make it easier to identify and fix the problems
in the native client.